Finish porting dungeons-1.21.11 utils/ and features/ to Kotlin
Converts the remaining Java files in fishmod/utils/{config/values,dungeon}
and fishmod/features/** (including croesus, dungeon, item, other subpackages)
to Kotlin, following the established conventions from the rest of the port:
objects with @JvmStatic/@JvmField for static/singleton classes, plain classes
with explicit accessor methods for former Java records, and Screen subclasses
for the GUI screens (FishModScreen, ItemCustomizeScreen, LootTrackerScreen,
DungeonWaypointListScreen, FishHudEditor).

Also fixes a handful of pre-existing Int/Float/Double argument mismatches and
a smart-cast issue in other already-converted Kotlin files that were
uncovered once the full utils/features Kotlin source set could compile
together (F7Huds, PuzzleDisplay, PartyFinderStats, EntityUtil, RunHistory,
HypixelApi).

+++ dungeons-1.21.11/src/main/kotlin/fishmod/utils/data/EntityUtil.kt
@@ -91,7 +91,7 @@ object EntityUtil {
 
     @JvmStatic
     fun getBox(entity: Entity): Box {
-        val tickProgress = MinecraftClient.getInstance().renderTickCounter.getTickProgress(false)
+        val tickProgress = MinecraftClient.getInstance().renderTickCounter.getTickProgress(false).toDouble()
 
         val pos = Misc.getPos(entity, tickProgress)
 
@@ -101,7 +101,7 @@ object EntityUtil {
 
     @JvmStatic
     fun getLerpedPos(entity: Entity): Vec3d {
-        val tickProgress = MinecraftClient.getInstance().renderTickCounter.getTickProgress(false)
+        val tickProgress = MinecraftClient.getInstance().renderTickCounter.getTickProgress(false).toDouble()
         return Misc.getPos(entity, tickProgress)
     }
 }
